requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
awswrangler==3.11.0
boto3==1.36.1
botocore==1.36.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickhouse-connect==0.8.15
colorama==0.4.6
frozenlist==1.5.0
greenlet==3.1.1
hypersync==0.8.5
idna==3.10
jmespath==1.0.1
lz4==4.4.3
minio==7.2.15
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
polars==1.19.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
SQLAlchemy==2.0.37
StrEnum==0.4.15
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
